refactor: extract shared code to stardb core package
Move reusable code from engine packages to stardb shared core:

- Interner (o11ytsdb → stardb): high-perf string→int hash table
- BackpressureController (o11ytsdb → stardb): async counting semaphore
- lowerBound/upperBound (o11ytsdb → stardb): BigInt64Array binary search
- Bloom filter (o11ytracesdb → stardb): BF8 with double hashing
- buildDictWithIndex (o11ytracesdb → stardb): frequency-sorted dictionary
- encodeAnyValue/decodeAnyValue (o11ytracesdb → stardb): binary AnyValue codec
- uint8IndexOf (o11ylogsdb → stardb): SIMD-accelerated byte search

Engine packages now re-export from stardb, preserving all public APIs.
All 1155 tests pass, typecheck clean.
