+++ packages/stardb/src/index.ts
@@ -4,7 +4,11 @@
 // `o11ylogsdb` and (in time) `o11ytsdb` and `o11ytracesdb` consume.
 
 export { anyValueEquals, anyValueToJson, findAttribute, jsonToAnyValue } from "./any-value.js";
+export { decodeAnyValue, encodeAnyValue, ValueTag } from "./any-value-binary.js";
+export { BackpressureController } from "./backpressure.js";
 export { ByteBuf, ByteReader } from "./binary.js";
+export { lowerBound, upperBound } from "./binary-search.js";
+export { bloomFromBase64, bloomMayContain, bloomToBase64, createBloomFilter } from "./bloom.js";
 export type { ChunkWireOptions } from "./chunk-wire.js";
 export { chunkWireSize, deserializeChunkWire, serializeChunkWire } from "./chunk-wire.js";
 export type { Codec, IntCodec, StringCodec } from "./codec.js";
@@ -17,6 +21,8 @@ export {
   rawInt64Codec,
   ZstdCodec,
 } from "./codec-baseline.js";
+export type { InternId } from "./interner.js";
+export { Interner } from "./interner.js";
 export { StreamRegistry } from "./stream.js";
 export type {
   AnyValue,
@@ -28,6 +34,7 @@ export type {
   StreamKey,
 } from "./types.js";
 export {
+  buildDictWithIndex,
   bytesEqual,
   bytesToHex,
   bytesToUuid,
@@ -35,5 +42,6 @@ export {
   hexToBytes,
   nowMillis,
   timeRangeOverlaps,
+  uint8IndexOf,
   uuidToBytes,
 } from "./utils.js";
