+++ packages/stardb/test/public-api.test.ts
@@ -10,6 +10,7 @@ import type {
   Codec,
   InstrumentationScope,
   IntCodec,
+  InternId,
   KeyValue,
   Resource,
   SeverityText,
@@ -24,6 +25,11 @@ import * as stardb from "../src/index.js";
 const RUNTIME_EXPORTS = [
   "anyValueEquals",
   "anyValueToJson",
+  "BackpressureController",
+  "bloomFromBase64",
+  "bloomMayContain",
+  "bloomToBase64",
+  "buildDictWithIndex",
   "bytesEqual",
   "bytesToHex",
   "bytesToUuid",
@@ -31,21 +37,29 @@ const RUNTIME_EXPORTS = [
   "ByteReader",
   "chunkWireSize",
   "CodecRegistry",
+  "createBloomFilter",
+  "decodeAnyValue",
   "defaultRegistry",
   "deserializeChunkWire",
+  "encodeAnyValue",
   "findAttribute",
   "fnv1aBytes",
   "GzipCodec",
   "hexToBytes",
+  "Interner",
   "jsonToAnyValue",
   "lengthPrefixStringCodec",
+  "lowerBound",
   "nowMillis",
   "rawCodec",
   "rawInt64Codec",
   "serializeChunkWire",
   "StreamRegistry",
   "timeRangeOverlaps",
+  "uint8IndexOf",
+  "upperBound",
   "uuidToBytes",
+  "ValueTag",
   "ZstdCodec",
 ] as const;
